+++ movie_rental_backend/src/main/java/com/rental_backend/repository/SuggestionRepository.java
@@ -15,8 +15,8 @@ public interface SuggestionRepository extends CrudRepository<Suggestion,Long> {
     List<Suggestion> findAll();
 
     @Query("select s.movie.title from Suggestion s, Customer c where s.primaryKey.msender_id = c.uId and s.primaryKey.msender_id = :senderId ")
-    List<Suggestion> findSuggestionBySenderId(@Param("senderId") Long senderId);
+    List<Suggestion> findSuggestionBySenderId(@Param("senderId") int senderId);
 
     @Query("select s.movie.title,s.suggestionSender.email,s.suggestionSender.name from Suggestion s, Customer c where s.primaryKey.mreceiver_id = c.uId and s.primaryKey.mreceiver_id = :receiverId ")
-    List<Suggestion> findSuggestionByReceiverId(@Param("receiverId") Long receiverId);
+    List<Suggestion> findSuggestionByReceiverId(@Param("receiverId") int receiverId);
 }
